feat: db to redis warming

+++ chat-service/src/main/java/com/pickpack/chatservice/dto/IsNewDto.java
@@ -8,12 +8,12 @@ import java.util.List;
 @Getter
 public class IsNewDto {
     private String lastMessage;
-    private int size;
+    private Long size;
 
     public static IsNewDto create(List<RedisChatMessage> list){
         IsNewDto isNewDto = new IsNewDto();
         isNewDto.lastMessage=list.get(list.size()-1).getMessage();
-        isNewDto.size=list.size();
+        isNewDto.size=(long)list.size();
         return isNewDto;
     }
 }
